Added card and deck structure for Issue #2
Removed ReturnDeckSize function and its associated test
Implemented card struct and deck type.
Implemented New() to fill a new deck with 52 cards: 4 suits, 13 cards each.
Implemented GetCardIndex function to return the index of card when supplied with a deck, a card value, and a suit string.
Implemented test for GetCardIndex.