Probability functions and helpers for issue #16
Added two helper functions: ValSortCardDes and ValSortCardAsc to sort an array of cards by value.

Added StraightCheck and tests:
Currrently just tests for a straight in an array of cards and returns whether or not it is a royal straight.

Included RemoveCards call to beginning of checks and removed it from associated tests.