Created deck probability helper function FindCardProb #16
FindCardProb takes a card array, an int array with needed values, a string with the needed suit or "" for none, and a number for how many more cards of the specified suit are needed for a flush. Statistics work is likely off at this point, but is close enough for basic functionality.